refactor(infra): reorganize sqlite driver and repository under db/sqlite/
