+++ packages/infrastructure/src/db/sqlite/sqlite-client.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest';
-import { db } from './db.js';
-import { documents } from './schema/documents.js';
+import { documents } from '../schema/documents.js';
+import { db } from './sqlite-client.js';
 
-describe('Database client & schema', () => {
+describe('SQLite client & schema', () => {
   it('should initialize db and bootstrap documents table', async () => {
     expect(db).toBeDefined();
