+++ packages/infrastructure/src/index.ts
@@ -1,6 +1,6 @@
-export * from './db/db.js';
 export * from './db/repositories/document-repository.js';
-export * from './db/repositories/sqlite-document-repository.js';
 export * from './db/schema/documents.js';
+export * from './db/sqlite/sqlite-client.js';
+export * from './db/sqlite/sqlite-document-repository.js';
 export * from './storage/local-storage.js';
 export * from './storage/storage-service.js';
